chore: updates per PR feedback
