+++ pkg/graph/schema/types/analyzed_audit.graphql
@@ -51,7 +51,7 @@ type AnalyzedPlanSections {
     updated: [TableName!]!
     readyForReview: [TableName!]!
     readyForClearance: [TableName!]!
-    dataExchangeApproachMarkedComplete: Boolean!
+    dataExchangeApproachMarkedComplete: Boolean
 }
 
 type AnalyzedModelLeads {
